move test utils to different folder to be able to use them both from jest tests and testcafe
