+++ frontend/benefit/applicant/jest.config.js
@@ -7,17 +7,16 @@ module.exports = {
     },
   },
   moduleNameMapper: {
-    ['^test-utils']: '<rootDir>/../../shared/test/test-utils',
     [`^shared\/(.*)$`]: '<rootDir>/../../shared/src/$1',
     [`benefit/applicant\/(.*)$`]: '<rootDir>/src/$1',
   },
   setupFilesAfterEnv: [
-    '<rootDir>/../../shared/test/setupTests.ts',
+    '<rootDir>/../../shared/src/__tests__/utils/setupTests.ts',
     '<rootDir>/test/i18n/i18n-test.ts',
   ],
   coveragePathIgnorePatterns: [
     '<rootDir>/src/pages/',
     '<rootDir>/../../shared/src/server/next-server.js',
-    '<rootDir>/../../shared/test/',
+    '<rootDir>/../../shared/src/test/',
   ],
 };
