+++ frontend/jest.config.js
@@ -2,7 +2,6 @@ module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'jsdom',
   moduleNameMapper: {
-    ['^test-utils']: '<rootDir>/shared/test/test-utils',
     [`^shared\/(.*)$`]: '<rootDir>/shared/src/$1',
     [`kesaseteli/employer\/(.*)$`]: '<rootDir>/kesaseteli/employer/src/$1',
     [`kesaseteli/handler\/(.*)$`]: '<rootDir>/kesaseteli/handler/src/$1',
@@ -10,7 +9,7 @@ module.exports = {
     [`benefit/handler\/(.*)$`]: '<rootDir>/benefit/handler/src/$1',
   },
   setupFilesAfterEnv: [
-    '<rootDir>/shared/test/setupTests.ts',
+    '<rootDir>/shared/src/__tests_/utils/setupTests.ts',
     '<rootDir>benefit/applicant/test/i18n/i18n-test.ts',
     '<rootDir>kesaseteli/employer/src/__tests__/utils/i18n/i18n-test.ts',
   ],
@@ -18,7 +17,7 @@ module.exports = {
     '<rootDir>/kesaseteli/employer/src/pages/',
     '<rootDir>/benefit/applicant/src/pages/',
     '<rootDir>/shared/src/server/next-server.js',
-    '<rootDir>/shared/test/',
+    '<rootDir>/shared/src/test/',
   ],
   collectCoverageFrom: ['./**/*.{ts,tsx,js,jsx}'],
   testPathIgnorePatterns: ['utils'],
