+++ frontend/kesaseteli/employer/jest.config.js
@@ -7,12 +7,11 @@ module.exports = {
     },
   },
   moduleNameMapper: {
-    ['^test-utils']: '<rootDir>/../../shared/test/test-utils',
     [`^shared\/(.*)$`]: '<rootDir>/../../shared/src/$1',
-    [`kesaseteli/employer\/(.*)$`]: '<rootDir>src/$1',
+    [`^kesaseteli/employer\/(.*)$`]: '<rootDir>src/$1',
   },
   setupFilesAfterEnv: [
-    '<rootDir>/../../shared/test/setupTests.ts',
+    '<rootDir>/../../shared/src/__tests__/utils/setupTests.ts',
     '<rootDir>src/__tests__/utils/i18n/i18n-test.ts',
   ],
   coveragePathIgnorePatterns: ['<rootDir>/kesaseteli/employer/src/pages/'],
