+++ frontend/kesaseteli/employer/src/__tests__/utils/backend/backend-nocks.ts
@@ -1,9 +1,9 @@
-import { fakeUser } from 'kesaseteli/employer/__tests__/utils/fake-objects';
 import {
   BackendEndpoint,
   getBackendDomain,
 } from 'kesaseteli/employer/backend-api/backend-api';
 import nock from 'nock';
+import { fakeUser } from 'shared/__tests__/utils/fake-objects';
 import Application from 'shared/types/employer-application';
 import User from 'shared/types/user';
