+++ frontend/kesaseteli/employer/src/__tests__/utils/components/render-component.tsx
@@ -3,8 +3,8 @@ import { NextRouter } from 'next/router';
 import React from 'react';
 import { QueryClient, QueryClientProvider } from 'react-query';
 import getDefaultReactQueryTestClient from 'shared/__tests__/utils/react-query/get-default-react-query-test-client';
+import { render, RenderResult } from 'shared/__tests__/utils/test-utils';
 import BackendAPIContext from 'shared/backend-api/BackendAPIContext';
-import { render, RenderResult } from 'test-utils';
 
 const renderComponent = (
   Component: JSX.Element,
