+++ frontend/kesaseteli/employer/src/__tests__/utils/components/render-page.tsx
@@ -6,12 +6,12 @@ import { NextPage } from 'next';
 import { NextRouter } from 'next/router';
 import React from 'react';
 import { QueryClient, QueryClientProvider } from 'react-query';
+import { render, RenderResult } from 'shared/__tests__/utils/test-utils';
 import BackendAPIContext from 'shared/backend-api/BackendAPIContext';
 import Content from 'shared/components/content/Content';
 import Layout from 'shared/components/layout/Layout';
 import theme from 'shared/styles/theme';
 import { ThemeProvider } from 'styled-components';
-import { render, RenderResult } from 'test-utils';
 
 const renderPage = (
   Page: NextPage,
