+++ frontend/shared/src/components/content/__tests__/Content.test.tsx
@@ -1,6 +1,6 @@
 import { axe } from 'jest-axe';
 import React from 'react';
-import { render } from 'test-utils';
+import { render } from 'shared/__tests__/utils/test-utils';
 
 import Content from '../Content';
