+++ frontend/shared/src/components/header/__tests__/Header.test.tsx
@@ -1,6 +1,6 @@
 import { axe } from 'jest-axe';
 import React from 'react';
-import { render } from 'test-utils';
+import { render } from 'shared/__tests__/utils/test-utils';
 
 import Header from '../Header';
