+++ frontend/shared/src/components/layout/__tests__/Layout.test.tsx
@@ -1,6 +1,6 @@
 import { axe } from 'jest-axe';
 import React from 'react';
-import { render } from 'test-utils';
+import { render } from 'shared/__tests__/utils/test-utils';
 
 import Layout from '../Layout';
