+++ frontend/shared/src/components/stepper/__tests__/Stepper.test.tsx
@@ -4,7 +4,7 @@ import { StepProps } from 'shared/components/stepper/Step';
 import Stepper from 'shared/components/stepper/Stepper';
 import theme from 'shared/styles/theme';
 import { ThemeProvider } from 'styled-components';
-import { render, screen } from 'test-utils';
+import { render, screen } from 'shared/__tests__/utils/test-utils';
 
 test('test for accessibility violations', async () => {
   const steps: StepProps[] = [{ title: 'Yritys' }];
