+++ frontend/shared/src/types/employer-application.d.ts
@@ -1,6 +1,6 @@
-import type Status from 'shared/types//application-status';
-import type Invoicer from 'shared/types//invoicer';
+import type Status from 'shared/types/application-status';
 import type Company from 'shared/types/company';
+import type Invoicer from 'shared/types/invoicer';
 
 type EmployerApplication = Invoicer & {
   id: string;
